+++ Modules/Microsoft365DSC/Modules/M365DSCDRGUtil.psm1
@@ -2711,7 +2711,6 @@ function Update-IntuneDeviceConfigurationPolicy
         [Array]
         $Settings
     )
-    $VerbosePreference = 'continue'
     try
     {
         $Uri = (Get-MSCloudLoginConnectionProfile -Workload MicrosoftGraph).ResourceUrl + "beta/deviceManagement/configurationPolicies/$DeviceConfigurationPolicyId"